vim: move file encoding to top to fix issues
In some cases (like a dev environment I just provisioned), `encoding`
is `latin1` by default.

This caused problems with the following line, which contains characters
outside of the Latin-1 set:

    set listchars=tab:▸\ ,eol:¬,trail:·,nbsp:·

Moving `set encoding=UTF-8` to the top of my `init.vim` fixed this. I
don't think this is an issue in NeoVim (which defaults to UTF-8), but
it was an issue in Vim 9.0.1499.